refactor: Extract interface for StorageHandler
